refactor: create separate components folder
