+++ src/components/guess-the-guild/components/BlankGuildCard.tsx
