+++ src/pages/guess-the-guild.tsx
@@ -1,7 +1,7 @@
 import { Center, Text } from "@chakra-ui/react"
 import { ThemeProvider, useThemeContext } from "components/[guild]/ThemeContext"
 import Layout from "components/common/Layout"
-import { GameMenu } from "components/guess-the-guild/GameMenu"
+import { GameMenu } from "components/guess-the-guild/components/GameMenu"
 import { GuessGuildByLogo } from "components/guess-the-guild/gamemodes/GuessGuildByLogo"
 import { PairLogosToGuilds } from "components/guess-the-guild/gamemodes/PairLogosToGuilds"
 import { ReactNode, useState } from "react"
